requirements.txt
Django==4.1.4
djangorestframework==3.13.1
gunicorn==20.1.0
redis==4.3.3
walrus==0.9.2
celery==5.2.7
requests==2.27.1
python-dotenv==0.19.2
pyproj==3.3.1
pandas==1.4.3
Shapely==1.8.1
python-jose==3.3.0
cryptography==38.0.3
pyjwt==2.4.0
jwcrypto==1.4.2
geojson==2.5.0
arrow==1.2.2
tldextract==3.2.1
rtree==1.0.0
psycopg2-binary==2.9.3
dacite==1.6.0
marshmallow==3.17.0